chore(riklet): add tests to iptable package
